+++ examples/raspberry_pi/canopen/canable/sdo/PMSM_Torque_Control_Encoder_based.py
@@ -4,7 +4,7 @@
 # Title: SoloPy
 # Author: SOLOMotorControllers
 # Date: 2025
-# Code version: 4.0.0
+# Code version: 4.1.0
 # Availability: https://github.com/Solo-FL/SoloPy/tree/main/SoloPy
 # This Library is made by SOLOMotorControllers.COM
 # please visit:  https://www.SOLOMotorControllers.com/
@@ -19,7 +19,7 @@ import time
 # to know more please read: https:#www.solomotorcontrollers.com/how-to-connect-calibrate-incremental-encoder-with-solo/
 
 #tested with CANable V2.0 USB-can coverter, after connection the usb of CANable check cmd:" dmesg | tail -n 20 " show usage of /dev/ttyACM0
-mySolo = solo.SOLOMotorControllersCanopen(0, solo.CanBusBaudRate.RATE_1000, solo.CanCommunicationInterface.CANABLE, 'slcan', "/dev/ttyACM0")
+mySolo = solo.SoloMotorControllersCanopen(0, solo.CanBusBaudRate.RATE_1000, solo.CanCommunicationInterface.CANABLE, 'slcan', "/dev/ttyACM0")
 
 
 # Desired Switching or PWM Frequency at Output