+++ examples/windows/canopen/canable/sdo/BLDC_Drone_RC_Car_Sensorless.py
@@ -4,7 +4,7 @@
 # Title: SoloPy
 # Author: SOLOMotorControllers
 # Date: 2025
-# Code version: 4.0.0
+# Code version: 4.1.0
 # Availability: https://github.com/Solo-FL/SoloPy/tree/main/SoloPy
 # This Library is made by SOLOMotorControllers.COM
 # please visit:  https://www.SOLOMotorControllers.com/
@@ -17,7 +17,7 @@ import SoloPy as solo
 import time
 
 #tested with CANable V2.0 USB-can coverter, device manager show CANable at COM69 
-mySolo = solo.SOLOMotorControllersCanopen(0, solo.CanBusBaudRate.RATE_1000, solo.CanCommunicationInterface.CANABLE, 'slcan', 'COM69')
+mySolo = solo.SoloMotorControllersCanopen(0, solo.CanBusBaudRate.RATE_1000, solo.CanCommunicationInterface.CANABLE, 'slcan', 'COM69')
 
 # Desired Switching or PWM Frequency at Output
 pwmFrequency = 79